Corrigindo codigos

+++ Projeto2/Models/Produto.cs
@@ -5,6 +5,6 @@
         public int Id { get; set; }
         public string? Nome { get; set; }
         public string? Descricao { get; set; }
-        public int? Preco { get; set; }
+        public decimal? Preco { get; set; }
     }
 }
